chore: 나의 스터디 페이지 모달 조건부 렌더링 방식으로 변경
