--- apps/client/app/(afterLogin)/my-study/attendance-complete/page.tsx
@@ -1,8 +1,0 @@
-import { routePath } from "constants/routePath";
-import { redirect } from "next/navigation";
-
-const AttendanceCompletePage = () => {
-  return redirect(routePath["my-study"]);
-};
-
-export default AttendanceCompletePage;
